llamacpp/app/requirements.txt
pyyaml
flask
nvidia-ml-py
